+++ apps/starter/CHANGELOG.md
@@ -1,5 +1,12 @@
 # stoker
 
+## 0.4.80
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/cli@0.5.53
+
 ## 0.4.79
 
 ### Patch Changes